refactor(network): clarify transport and upstream module layout
- Split upstream runtime code into focused modules for traits, pooled execution, bootstrap refresh, and bootstrap pool factories.
- Rename pool and transport modules to concise protocol and strategy names while preserving public transport compatibility re-exports.
- Keep existing upstream public re-exports intact so callers continue using the same upstream surface.
