+++ src/infra/network/upstream/builder.rs
@@ -4,6 +4,9 @@
 use tracing::debug;
 
 use crate::infra::error::Result;
+use crate::infra::network::upstream::bootstrap::{
+    BootstrapUdpTruncatedUpstream, BootstrapUpstream,
+};
 use crate::infra::network::upstream::config::{ConnectionInfo, ConnectionType, UpstreamConfig};
 #[cfg(feature = "upstream-doh")]
 use crate::infra::network::upstream::conn::{H2Connection, H2ConnectionBuilder};
@@ -14,13 +17,11 @@ use crate::infra::network::upstream::conn::{QuicConnection, QuicConnectionBuilde
 use crate::infra::network::upstream::conn::{
     TcpConnection, TcpConnectionBuilder, UdpConnectionBuilder,
 };
-use crate::infra::network::upstream::pool::pool_pipeline::PipelinePool;
-use crate::infra::network::upstream::pool::pool_reuse::ReusePool;
+use crate::infra::network::upstream::pool::pipeline::PipelinePool;
+use crate::infra::network::upstream::pool::reuse::ReusePool;
 use crate::infra::network::upstream::pool::{Connection, ConnectionBuilder, QueryTimeoutPolicy};
-use crate::infra::network::upstream::resolver::{
-    BootstrapUdpTruncatedUpstream, BootstrapUpstream, PooledUpstream, UdpTruncatedUpstream,
-    Upstream,
-};
+use crate::infra::network::upstream::pooled::{PooledUpstream, UdpTruncatedUpstream};
+use crate::infra::network::upstream::traits::Upstream;
 
 /// Builder for creating upstream instances
 pub struct UpstreamBuilder;
